requirements.txt
streamlit>=1.25
google-genai>=1.37.0
python-dotenv
